[IMP] res_currency_cr_adapter
