+++ res_currency_cr_adapter/i18n/es_CR.po
@@ -4,16 +4,33 @@
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 14.0-20220204\n"
+"Project-Id-Version: Odoo Server 15.0\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2022-04-11 20:53+0000\n"
-"PO-Revision-Date: 2022-04-11 20:53+0000\n"
+"POT-Creation-Date: 2022-09-25 01:05+0000\n"
+"PO-Revision-Date: 2022-09-25 01:05+0000\n"
 "Last-Translator: Norlan Ruiz<ruiznorlan@gmail.com>\n"
 "Language-Team: \n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
 "Content-Transfer-Encoding: \n"
 "Plural-Forms: \n"
+
+#. module: res_currency_cr_adapter
+#: code:addons/res_currency_cr_adapter/models/res_currency_rate.py:0
+#, python-format
+msgid "%(company_currency_name)s per %(rate_currency_name)s - Buy"
+msgstr "%(company_currency_name)s per %(rate_currency_name)s - Compra"
+
+#. module: res_currency_cr_adapter
+#: code:addons/res_currency_cr_adapter/models/res_currency_rate.py:0
+#, python-format
+msgid "%(rate_currency_name)s per %(company_currency_name)s - Buy"
+msgstr "%(rate_currency_name)s per %(company_currency_name)s - Compra"
+
+#. module: res_currency_cr_adapter
+#: model_terms:ir.ui.view,arch_db:res_currency_cr_adapter.res_config_settings_crc_view_form
+msgid "<span class=\"o_form_label\">Costa Rica Currency Adapter</span>"
+msgstr "<span class=\"o_form_label\">Adaptador de moneda de Costa Rica</span>"
 
 #. module: res_currency_cr_adapter
 #: model:ir.model.fields.selection,name:res_currency_cr_adapter.selection__res_config_settings__exchange_source__bccr
@@ -29,11 +46,6 @@ msgstr "Token para BCCR"
 #: model:ir.model.fields,field_description:res_currency_cr_adapter.field_res_config_settings__bccr_username
 msgid "BCCR username"
 msgstr "Nombre de Usuario para BCCR"
-
-#. module: res_currency_cr_adapter
-#: model:ir.model.fields,field_description:res_currency_cr_adapter.field_res_currency_rate__rate_2
-msgid "Buying Rate"
-msgstr "Tipo de cambio Compra"
 
 #. module: res_currency_cr_adapter
 #: model:ir.model.fields,field_description:res_currency_cr_adapter.field_res_currency_rate__original_rate_2
@@ -80,13 +92,6 @@ msgid "Disabled"
 msgstr "Desactivado"
 
 #. module: res_currency_cr_adapter
-#: model:ir.model.fields,field_description:res_currency_cr_adapter.field_res_config_settings__display_name
-#: model:ir.model.fields,field_description:res_currency_cr_adapter.field_res_currency__display_name
-#: model:ir.model.fields,field_description:res_currency_cr_adapter.field_res_currency_rate__display_name
-msgid "Display Name"
-msgstr "Mostrar nombre"
-
-#. module: res_currency_cr_adapter
 #: model_terms:ir.ui.view,arch_db:res_currency_cr_adapter.res_config_settings_crc_view_form
 msgid "Email registered in the BCCR"
 msgstr "Correo electrónico registrado en el BCCR"
@@ -97,33 +102,19 @@ msgid "Exchange Source"
 msgstr "Origen del tipo de Cambio"
 
 #. module: res_currency_cr_adapter
+#: model_terms:ir.ui.view,arch_db:res_currency_cr_adapter.res_config_settings_crc_view_form
+msgid "Exchange rate origin"
+msgstr "Origen del tipo de cambio"
+
+#. module: res_currency_cr_adapter
 #: model:ir.model.fields.selection,name:res_currency_cr_adapter.selection__res_config_settings__exchange_source__hacienda
 msgid "Hacienda"
 msgstr ""
 
 #. module: res_currency_cr_adapter
-#: model:ir.model.fields,field_description:res_currency_cr_adapter.field_res_config_settings__id
-#: model:ir.model.fields,field_description:res_currency_cr_adapter.field_res_currency__id
-#: model:ir.model.fields,field_description:res_currency_cr_adapter.field_res_currency_rate__id
-msgid "ID"
-msgstr "Identificación"
-
-#. module: res_currency_cr_adapter
-#: model:ir.model.fields,field_description:res_currency_cr_adapter.field_res_config_settings____last_update
-#: model:ir.model.fields,field_description:res_currency_cr_adapter.field_res_currency____last_update
-#: model:ir.model.fields,field_description:res_currency_cr_adapter.field_res_currency_rate____last_update
-msgid "Last Modified on"
-msgstr "Última modificación el"
-
-#. module: res_currency_cr_adapter
-#: model_terms:ir.ui.view,arch_db:res_currency_cr_adapter.res_config_settings_crc_view_form
-msgid "Origen del tipo de cambio"
-msgstr ""
-
-#. module: res_currency_cr_adapter
-#: model_terms:ir.ui.view,arch_db:res_currency_cr_adapter.view_currency_form
-msgid "Rates"
-msgstr "Tarifas"
+#: model:ir.model.fields,field_description:res_currency_cr_adapter.field_res_currency_rate__inverse_original_rate_2
+msgid "Inverse Original Rate 2"
+msgstr "Tasa original inversa 2"
 
 #. module: res_currency_cr_adapter
 #: model:ir.model.fields,field_description:res_currency_cr_adapter.field_res_currency_rate__rate
@@ -136,6 +127,11 @@ msgid "Selling Rate in Costa Rica"
 msgstr "Tipo de cambio Venta en Costa Rica"
 
 #. module: res_currency_cr_adapter
+#: model:ir.model.fields,field_description:res_currency_cr_adapter.field_res_currency_rate__rate_2
+msgid "Technical Rate - Buy"
+msgstr "Tasa Técnica - Compra"
+
+#. module: res_currency_cr_adapter
 #: model:ir.model.fields,help:res_currency_cr_adapter.field_res_currency_rate__original_rate_2
 msgid "The buying exchange rate from CRC to USD as it is send from BCCR"
 msgstr "El tipo de cambio de compra de CRC a USD tal como se envía desde BCCR"
@@ -146,9 +142,32 @@ msgid "The buying rate of the currency to the currency of rate 1."
 msgstr "El tipo de cambio de compra de la moneda a la moneda de tipo 1."
 
 #. module: res_currency_cr_adapter
+#: code:addons/res_currency_cr_adapter/models/res_currency_rate.py:0
+#, python-format
+msgid ""
+"The date for the current rate is empty.\n"
+"Please set it."
+msgstr ""
+"La fecha de la tasa actual no puede estar vacía.\n"
+"Por favor establecerla."
+
+#. module: res_currency_cr_adapter
+#: code:addons/res_currency_cr_adapter/models/res_currency_rate.py:0
+#, python-format
+msgid ""
+"The new rate is quite far from the previous rate.\n"
+"Incorrect currency rates may cause critical problems, make sure the rate is correct !"
+msgstr ""
+
+#. module: res_currency_cr_adapter
 #: model:ir.model.fields,help:res_currency_cr_adapter.field_res_currency_rate__rate
 msgid "The rate of the currency to the currency of rate 1"
 msgstr "La tasa de cambio de la moneda respecto a la moneda de tasa 1"
+
+#. module: res_currency_cr_adapter
+#: model:ir.model.fields,help:res_currency_cr_adapter.field_res_currency_rate__inverse_original_rate_2
+msgid "The rate of the currency to the currency of rate 1 "
+msgstr "La tasa de la moneda a la moneda de la tasa 1"
 
 #. module: res_currency_cr_adapter
 #: model:ir.model.fields,help:res_currency_cr_adapter.field_res_currency__rate
@@ -172,6 +191,11 @@ msgstr "Actualizar"
 
 #. module: res_currency_cr_adapter
 #: model_terms:ir.ui.view,arch_db:res_currency_cr_adapter.res_config_settings_crc_view_form
+msgid "Update the dollar exchange rate from Ministerio de Hacienda or BCCR"
+msgstr "Actualice el tipo de cambio del dólar desde el Ministerio de Hacienda o BCCR"
+
+#. module: res_currency_cr_adapter
+#: model_terms:ir.ui.view,arch_db:res_currency_cr_adapter.res_config_settings_crc_view_form
 msgid "User name registered in the BCCR"
 msgstr "Nombre de usuario registrado en el BCCR"
 
@@ -179,6 +203,12 @@ msgstr "Nombre de usuario registrado en el BCCR"
 #: model_terms:ir.ui.view,arch_db:res_currency_cr_adapter.res_config_settings_crc_view_form
 msgid "Username registered in the BCCR"
 msgstr "Nombre de usuario registrado en el BCCR"
+
+#. module: res_currency_cr_adapter
+#: code:addons/res_currency_cr_adapter/models/res_currency_rate.py:0
+#, python-format
+msgid "Warning for %s"
+msgstr "Alerta par %s"
 
 #. module: res_currency_cr_adapter
 #: model:ir.model.fields,field_description:res_currency_cr_adapter.field_res_config_settings__bccr_email
