+++ res_currency_cr_adapter/models/__init__.py
@@ -1,5 +1,6 @@
 # copyright  2018 Carlos Wong, Akurey SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from . import res_currency
 from . import res_config_settings
+from . import res_currency
+from . import res_currency_rate
